updated commit

+++ src/data/repositries/Items.java
@@ -4,7 +4,6 @@ import data.models.Item;
 
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collection;
 
 public class Items {
 
@@ -15,7 +14,7 @@ public class Items {
        items = new ArrayList<>();
    }
 
-    public long count  () {
+    public int count  () {
         return items.size();
     }
